requirements.txt
streamlit
pandas
scikit-learn
psycopg2-binary
